# longaeva/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AttributionRead": {
        "properties": {
          "baseline_run_id": {
            "format": "uuid",
            "title": "Baseline Run Id",
            "type": "string"
          },
          "conditional_on_model": {
            "title": "Conditional On Model",
            "type": "boolean"
          },
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/ContributionRead"
            },
            "title": "Contributions",
            "type": "array"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricAttributionRead"
            },
            "title": "Metrics",
            "type": "array"
          },
          "order_note": {
            "title": "Order Note",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "sensitivity": {
            "items": {
              "$ref": "#/components/schemas/SensitivityRead"
            },
            "title": "Sensitivity",
            "type": "array"
          },
          "sensitivity_metric": {
            "title": "Sensitivity Metric",
            "type": "string"
          },
          "sequential_order": {
            "items": {
              "type": "string"
            },
            "title": "Sequential Order",
            "type": "array"
          },
          "verification": {
            "items": {
              "type": "string"
            },
            "title": "Verification",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "baseline_run_id",
          "conditional_on_model",
          "order_note",
          "sequential_order",
          "metrics",
          "contributions",
          "sensitivity",
          "sensitivity_metric",
          "verification"
        ],
        "title": "AttributionRead",
        "type": "object"
      },
...
      "ComparisonRead": {
        "properties": {
          "baseline_run_id": {
            "format": "uuid",
            "title": "Baseline Run Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ComparisonRow"
            },
            "title": "Items",
            "type": "array"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "baseline_run_id",
          "seed",
          "items"
        ],
        "title": "ComparisonRead",
        "type": "object"
      },
      "ComparisonRow": {
        "properties": {
          "baseline_mean": {
            "title": "Baseline Mean",
            "type": "number"
          },
          "difference_mean": {
            "title": "Difference Mean",
            "type": "number"
          },
          "difference_std": {
            "title": "Difference Std",
            "type": "number"
          },
          "difference_std_error": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difference Std Error"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "period_label": {
            "title": "Period Label",
            "type": "string"
          },
          "quantile_std_errors": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Quantile Std Errors",
            "type": "object"
          },
          "quantiles": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Quantiles",
            "type": "object"
          },
          "quarter_index": {
            "title": "Quarter Index",
            "type": "integer"
          },
          "variant_mean": {
            "title": "Variant Mean",
            "type": "number"
          }
        },
        "required": [
          "metric",
          "quarter_index",
          "period_label",
          "baseline_mean",
          "variant_mean",
          "difference_mean",
          "difference_std"
        ],
        "title": "ComparisonRow",
        "type": "object"
      },
...
      "ContributionRead": {
        "properties": {
          "after": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "After"
          },
          "before": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before"
          },
          "by_metric": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MetricContributionRead"
            },
            "title": "By Metric",
            "type": "object"
          },
          "intervention": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intervention"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "enum": [
              "parameter",
              "intervention"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "observation_ids": {
            "items": {
              "type": "string"
            },
            "title": "Observation Ids",
            "type": "array"
          },
          "rule_ids": {
            "items": {
              "type": "string"
            },
            "title": "Rule Ids",
            "type": "array"
          },
          "size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          }
        },
        "required": [
          "kind",
          "name",
          "key",
          "label",
          "by_metric"
        ],
        "title": "ContributionRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetricAttributionRead": {
        "properties": {
          "horizon_total": {
            "title": "Horizon Total",
            "type": "number"
          },
          "joint_residual": {
            "items": {
              "$ref": "#/components/schemas/QuarterEffectRead"
            },
            "title": "Joint Residual",
            "type": "array"
          },
          "joint_residual_horizon": {
            "title": "Joint Residual Horizon",
            "type": "number"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "total": {
            "items": {
              "$ref": "#/components/schemas/QuarterEffectRead"
            },
            "title": "Total",
            "type": "array"
          }
        },
        "required": [
          "metric",
          "total",
          "horizon_total",
          "joint_residual",
          "joint_residual_horizon"
        ],
        "title": "MetricAttributionRead",
        "type": "object"
      },
...
      "MixShiftIntervention": {
        "additionalProperties": false,
        "properties": {
          "cross_border_change": {
            "exclusiveMinimum": -1.0,
            "maximum": 3.0,
            "title": "Cross Border Change",
            "type": "number"
          },
          "start_quarter": {
            "default": 1,
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Start Quarter",
            "type": "integer"
          },
          "type": {
            "const": "mix_shift_conserving_total",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "cross_border_change"
        ],
        "title": "MixShiftIntervention",
        "type": "object"
      },
...
      "PairRunsCreate": {
        "additionalProperties": false,
        "properties": {
          "baseline_scenario_id": {
            "format": "uuid",
            "title": "Baseline Scenario Id",
            "type": "string"
          },
          "cutoff_ts": {
            "format": "date-time",
            "title": "Cutoff Ts",
            "type": "string"
          },
          "n_paths": {
            "default": 5000,
            "maximum": 50000.0,
            "minimum": 1.0,
            "title": "N Paths",
            "type": "integer"
          },
          "n_quarters": {
            "default": 4,
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "N Quarters",
            "type": "integer"
          },
          "scenario_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 8,
            "minItems": 2,
            "title": "Scenario Ids",
            "type": "array"
          },
          "seed": {
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          },
          "switches": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Switches",
            "type": "object"
          }
        },
        "required": [
          "scenario_ids",
          "baseline_scenario_id",
          "cutoff_ts",
          "seed"
        ],
        "title": "PairRunsCreate",
        "type": "object"
      },
...
      "ParameterOverride": {
        "additionalProperties": false,
        "properties": {
          "rationale": {
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "value",
          "rationale"
        ],
        "title": "ParameterOverride",
        "type": "object"
      },
...
      "RunRead": {
        "properties": {
          "baseline_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Run Id"
          },
          "code_version": {
            "title": "Code Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cutoff_ts": {
            "format": "date-time",
            "title": "Cutoff Ts",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interventions": {
            "items": {},
            "title": "Interventions",
            "type": "array"
          },
          "interventions_hash": {
            "title": "Interventions Hash",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "lib_versions": {
            "additionalProperties": true,
            "title": "Lib Versions",
            "type": "object"
          },
          "n_paths": {
            "title": "N Paths",
            "type": "integer"
          },
          "n_quarters": {
            "title": "N Quarters",
            "type": "integer"
          },
          "origin_label": {
            "title": "Origin Label",
            "type": "string"
          },
          "outputs_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputs Hash"
          },
          "outputs_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputs Path"
          },
          "parameter_set_hash": {
            "title": "Parameter Set Hash",
            "type": "string"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "source_manifest": {
            "items": {},
            "title": "Source Manifest",
            "type": "array"
          },
          "source_manifest_hash": {
            "title": "Source Manifest Hash",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "starting_state_hash": {
            "title": "Starting State Hash",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "switches": {
            "additionalProperties": true,
            "title": "Switches",
            "type": "object"
          }
        },
        "required": [
          "id",
          "scenario_id",
          "cutoff_ts",
          "origin_label",
          "n_quarters",
          "source_manifest",
          "source_manifest_hash",
          "parameter_set_hash",
          "starting_state_hash",
          "code_version",
          "seed",
          "n_paths",
          "switches",
          "interventions_hash",
          "lib_versions",
          "status",
          "created_at"
        ],
        "title": "RunRead",
        "type": "object"
      },
...
      "ScenarioCreate": {
        "additionalProperties": false,
        "properties": {
          "company": {
            "minLength": 1,
            "title": "Company",
            "type": "string"
          },
          "interventions": {
            "items": {
              "discriminator": {
                "mapping": {
                  "mix_shift_conserving_total": "#/components/schemas/MixShiftIntervention",
                  "total_spend_reduction": "#/components/schemas/SpendReductionIntervention"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/MixShiftIntervention"
                },
                {
                  "$ref": "#/components/schemas/SpendReductionIntervention"
                }
              ]
            },
            "title": "Interventions",
            "type": "array"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "pair_group_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pair Group Id"
          },
          "parameter_overrides": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ParameterOverride"
            },
            "title": "Parameter Overrides",
            "type": "object"
          },
          "parameter_set_id": {
            "format": "uuid",
            "title": "Parameter Set Id",
            "type": "string"
          }
        },
        "required": [
          "company",
          "name",
          "parameter_set_id"
        ],
        "title": "ScenarioCreate",
        "type": "object"
      },
      "ScenarioRead": {
        "properties": {
          "company": {
            "title": "Company",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interventions": {
            "items": {},
            "title": "Interventions",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pair_group_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pair Group Id"
          },
          "parameter_set_id": {
            "format": "uuid",
            "title": "Parameter Set Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "company",
          "name",
          "parameter_set_id",
          "interventions",
          "created_at"
        ],
        "title": "ScenarioRead",
        "type": "object"
      },
...
      "SensitivityRead": {
        "properties": {
          "flag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flag"
          },
          "high_mean": {
            "title": "High Mean",
            "type": "number"
          },
          "low_mean": {
            "title": "Low Mean",
            "type": "number"
          },
          "normalized_sensitivity": {
            "title": "Normalized Sensitivity",
            "type": "number"
          },
          "observation_ids": {
            "items": {
              "type": "string"
            },
            "title": "Observation Ids",
            "type": "array"
          },
          "parameter": {
            "title": "Parameter",
            "type": "string"
          },
          "range_high": {
            "title": "Range High",
            "type": "number"
          },
          "range_low": {
            "title": "Range Low",
            "type": "number"
          },
          "rule_ids": {
            "items": {
              "type": "string"
            },
            "title": "Rule Ids",
            "type": "array"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          },
          "support_score": {
            "title": "Support Score",
            "type": "number"
          },
          "swing": {
            "title": "Swing",
            "type": "number"
          }
        },
        "required": [
          "parameter",
          "range_low",
          "range_high",
          "low_mean",
          "high_mean",
          "swing",
          "normalized_sensitivity",
          "support_score"
        ],
        "title": "SensitivityRead",
        "type": "object"
      },
...
      "SpendReductionIntervention": {
        "additionalProperties": false,
        "properties": {
          "reduction": {
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Reduction",
            "type": "number"
          },
          "start_quarter": {
            "default": 1,
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Start Quarter",
            "type": "integer"
          },
          "type": {
            "const": "total_spend_reduction",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "reduction"
        ],
        "title": "SpendReductionIntervention",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Standalone Visa business-simulation API (hackathon).",
    "title": "Longaeva",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/scenarios": {
      "get": {
        "operationId": "list_scenarios_scenarios_get",
        "parameters": [
          {
            "in": "query",
            "name": "company",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioRead"
                  },
                  "title": "Response List Scenarios Scenarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Scenarios",
        "tags": [
          "scenarios"
        ]
      },
      "post": {
        "operationId": "post_scenario_scenarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Scenario",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/scenarios/attribution": {
      "get": {
        "operationId": "get_attribution_scenarios_attribution_get",
        "parameters": [
          {
            "in": "query",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "baseline_run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Baseline Run Id"
            }
          },
          {
            "in": "query",
            "name": "metric",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Metric"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttributionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Attribution",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/scenarios/comparison": {
      "get": {
        "operationId": "get_comparison_scenarios_comparison_get",
        "parameters": [
          {
            "in": "query",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "baseline_run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Baseline Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Comparison",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/scenarios/pair-runs": {
      "post": {
        "operationId": "post_pair_runs_scenarios_pair_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairRunsCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunRead"
                  },
                  "title": "Response Post Pair Runs Scenarios Pair Runs Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Pair Runs",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/scenarios/{scenario_id}": {
      "get": {
        "operationId": "get_scenario_scenarios__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scenario",
        "tags": [
          "scenarios"
        ]
      }
    },
...
  }
}